refactor: organize bot services
