+++ api/src/components/translations/translations.module.ts
@@ -14,10 +14,11 @@ import { WordToPhraseTranslationsService } from './word-to-phrase-translations.s
 import { PhraseToPhraseTranslationsService } from './phrase-to-phrase-translations.service';
 import { PhraseToWordTranslationsService } from './phrase-to-word-translations.service';
 import { TranslationsService } from './translations.service';
-import { GoogleTranslateService } from './google-translate.service';
+import { GoogleTranslateService } from './translator-bots/google-translate.service';
 
 import { WordToWordTranslationRepository } from './word-to-word-translation.repository';
-import { LiltTranslateService } from './lilt-translate.service';
+import { LiltTranslateService } from './translator-bots/lilt-translate.service';
+import { AiTranslationsService } from './translator-bots/ai-translations.service';
 
 @Module({
   imports: [
@@ -38,6 +39,7 @@ import { LiltTranslateService } from './lilt-translate.service';
     GoogleTranslateService,
     LiltTranslateService,
     TranslationsService,
+    AiTranslationsService,
   ],
   exports: [
     WordToWordTranslationsService,
@@ -47,6 +49,7 @@ import { LiltTranslateService } from './lilt-translate.service';
     GoogleTranslateService,
     LiltTranslateService,
     TranslationsService,
+    AiTranslationsService,
   ],
 })
 export class TranslationsModule {}
